Work around bugs in gcc and glibc to fix test malloc
When building the tests with -O2, gcc makes a mess of optimizing our
custom test malloc (falloc.c).  gcc assumes any function called malloc
is builtin malloc, so we need to build with

        -fno-builtin-malloc -fno-builtin-calloc

Also, we need to make our lock variable volatile otherwise gcc will
omit assigning to it before calling dlsym. This seems to be a problem
with the dlsym() __THROW declaration in glibc. Works fine on NetBSD.

Signed-off-by: Brett A C Sheffield <bacs@librecast.net>