lc_socket_bind() - filter packets by interface
When a socket is bound to a specific interface we only want to receive
packets on that interface. If we bind to an interface-specific address,
we will get no multicast packets. If bound to either INADDR_ANY or the
multicast group address, we receive packets on all interfaces. So, we
need to filter by extracting the receiving interface from ancillary data