lc_socket_send() - send to all channels bound to socket
- lc_socket_send()

A unix socket(2) can only be bound to an address once, whereas a
Librecast Socket can be bound to multiple channels.  On first call to
lc_channel_bind() we call bind(2) and set socket options. On subsequent
calls we skip these steps and just mark the channel as bound to the
socket.